updated distance to double

+++ Backend/src/main/java/at/fhtw/swen2_tourplanner/backend/tour/dto/TourDTO.java
@@ -27,7 +27,7 @@ public class TourDTO {
     @JsonProperty("transport_type")
     private int transportType;
     @JsonProperty("tour_distance")
-    private long tourDistance;
+    private double tourDistance;
     @JsonProperty("estimated_time")
     @JsonFormat(pattern = "HH:mm:ss")
     private LocalTime estimatedTime;
